# internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: 社内OSSカタログ / 納品用 OSS 管理 API
  version: 1.0.0-draft2
  description: |
    社内で利用・納品対象となる OSS を一元管理し、プロジェクト毎の利用状況と納品用一覧（SPDX / CSV 等エクスポート）を生成するための API 初稿 (Phase 1)。

    **本稿の位置づけ**
    - Phase 1 対象: OSSコンポーネント/バージョン CRUD, タグ, プロジェクト, プロジェクト利用 (Usage), スコープ判定, ポリシー参照, 監査ログ最小, エクスポート(プレースホルダ)
    - 未実装(将来): 脆弱性, SBOM Import, NOTICE 生成, ライセンス全文管理, **JWT リフレッシュ**, 監査詳細検索高度化

    **表記**
    - ID は原則 UUID v4 文字列
    - 日本語説明は実装後英語化検討
servers:
  - url: https://api.example.local/api/v1
    description: 本番 (社内)
  - url: https://dev.api.example.local/api/v1
    description: 開発環境
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 互換用途の Basic 認証（将来廃止予定）
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: アクセストークン用 JWT（リフレッシュなし）
  parameters:
    PageParam:
      name: page
      in: query
      description: 1 始まりのページ番号
      schema: { type: integer, minimum: 1, default: 1 }
    SizeParam:
      name: size
      in: query
      description: 1ページ件数 (最大 200)
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    SortParam:
      name: sort
      in: query
      description: "ソート指定 (例: name,asc / createdAt,desc)"
      schema: { type: string }
  responses:
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: 認証失敗
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Layer:
      type: string
      enum:
        [
          OS,
          RUNTIME,
          DB,
          MIDDLEWARE,
          LIB,
          FRAMEWORK,
          TOOL_BUILD,
          TOOL_DEV,
          TOOL_TEST,
          DOC,
          OTHER,
        ]
      description: OSS 技術レイヤ分類（OS=OS, LIB=ライブラリ 等）
      x-enumDescriptions:
        - オペレーティングシステム（再配布するイメージや基盤）
        - 実行ランタイム/言語実装（JVM, Python, Node.js 等）
        - データベースエンジン
        - 通信/メッセージング/キャッシュ等ミドルウェア
        - ライブラリ（アプリにリンク/取り込み）
        - アプリ/言語フレームワーク（Spring, Vue 等）
        - ビルド専用ツール（Maven, CMake 等）
        - 開発支援/IDE/静的解析など
        - テスト専用ツール（JUnit, Catch2 等）
        - ドキュメント生成/文書化ツール
        - 上記に当てはまらないその他
...
    ScopeStatus:
      type: string
      enum: [IN_SCOPE, OUT_SCOPE, REVIEW_NEEDED]
      description: 納品対象スコープ判定状態（IN_SCOPE=含む, OUT_SCOPE=除外, REVIEW_NEEDED=要判定）
      x-enumDescriptions:
        - 納品対象に含める
        - 納品対象から除外
        - 判定保留（要レビュー）
...
    ReviewStatus:
      type: string
      enum: [draft, verified]
      description: バージョンレビュー状態（draft=未承認, verified=確認済）
      x-enumDescriptions:
        - 未レビュー/暫定登録
        - レビュー済/確認済
...
    OssComponent:
      type: object
      description: OSS の論理的名称（バージョン共通情報）
      properties:
        id: { type: string, format: uuid, description: OSSコンポーネント ID }
        name:
          {
            type: string,
            description: 表示名 / ユニーク（大文字小文字区別ポリシーは実装で決定）,
          }
        normalizedName:
          { type: string, description: 検索用正規化名称（小文字化・記号除去） }
        homepageUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: 公式サイト URL,
          }
        repositoryUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: ソースリポジトリ（Git 等）URL,
          }
        description:
          { type: string, nullable: true, description: 簡易説明 / メモ }
        primaryLanguage:
          {
            type: string,
            nullable: true,
            description: "主言語（例: C, C++, Go, Java）",
          }
        layers:
          type: array
          description: 技術レイヤ分類（複数指定可）
          items: { $ref: "#/components/schemas/Layer" }
        defaultUsageRole:
          $ref: "#/components/schemas/UsageRole"
          nullable: true
          description: 新規 ProjectUsage 作成時の初期 usageRole 推奨値
        deprecated: { type: boolean, description: 非推奨フラグ（新規利用抑止） }
        createdAt: { type: string, format: date-time, description: 作成日時 }
        updatedAt: { type: string, format: date-time, description: 更新日時 }
        tags:
          type: array
          description: 付与タグ一覧
          items: { $ref: "#/components/schemas/Tag" }
      required: [id, name, deprecated, createdAt, updatedAt]
    OssComponentCreateRequest:
      type: object
      description: OSSコンポーネント作成リクエスト
      required: [name]
      properties:
        name: { type: string, description: OSS 表示名 }
        homepageUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: 公式サイト URL,
          }
        repositoryUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: リポジトリ URL,
          }
        description: { type: string, nullable: true, description: 説明 / 備考 }
        primaryLanguage: { type: string, nullable: true, description: 主言語 }
        layers:
          type: array
          description: 技術レイヤ配列
          items: { $ref: "#/components/schemas/Layer" }
        defaultUsageRole:
          $ref: "#/components/schemas/UsageRole"
          nullable: true
          description: デフォルト usageRole
        tagIds:
          type: array
          description: 初期付与するタグ ID 群
          items: { type: string, format: uuid }
    OssComponentUpdateRequest:
      type: object
      description: OSSコンポーネント更新リクエスト（部分）
      properties:
        name: { type: string, description: "新しい名称" }
        homepageUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: "公式サイト URL",
          }
        repositoryUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: "リポジトリ URL",
          }
        description:
          { type: string, nullable: true, description: "説明 / 備考" }
        primaryLanguage: { type: string, nullable: true, description: "主言語" }
        layers:
          type: array
          description: 技術レイヤ配列
          items: { $ref: "#/components/schemas/Layer" }
        defaultUsageRole:
          $ref: "#/components/schemas/UsageRole"
          nullable: true
          description: "デフォルト usageRole"
        deprecated: { type: boolean, description: "非推奨フラグ" }
        tagIds:
          type: array
          description: 置換後のタグ ID 配列
          items: { type: string, format: uuid }
    OssVersion:
      type: object
      description: 個別バージョン情報
      properties:
        id: { type: string, format: uuid, description: "バージョン ID" }
        ossId:
          { type: string, format: uuid, description: "親 OSSコンポーネント ID" }
        version:
          { type: string, description: "バージョン文字列 (SemVer 互換推奨)" }
        releaseDate:
          {
            type: string,
            format: date,
            nullable: true,
            description: "上流リリース日",
          }
        licenseExpressionRaw:
          {
            type: string,
            nullable: true,
            description: "SPDX 互換ライセンス式（上流取得時点）",
          }
        licenseConcluded:
          {
            type: string,
            nullable: true,
            description: "社内審査確定ライセンス式",
          }
        purl:
          {
            type: string,
            nullable: true,
            description: "package-url (例: pkg:maven/...)",
          }
        cpeList:
          type: array
          description: CPE 文字列配列（脆弱性紐付け用）
          items: { type: string }
        hashSha256:
          type: string
          pattern: "^[A-Fa-f0-9]{64}$"
          nullable: true
          description: "配布アーカイブ等の SHA-256 ハッシュ"
        modified: { type: boolean, description: "社内改変有無" }
        modificationDescription:
          { type: string, nullable: true, description: "改変内容概要" }
        reviewStatus:
          {
            $ref: "#/components/schemas/ReviewStatus",
            description: "レビュー状態",
          }
        lastReviewedAt:
          {
            type: string,
            format: date-time,
            nullable: true,
            description: "最終レビュー日時",
          }
        scopeStatus:
          {
            $ref: "#/components/schemas/ScopeStatus",
            description: "バージョン単位の推奨スコープ判定（参考）",
          }
        supplierType:
          {
            $ref: "#/components/schemas/SupplierType",
            nullable: true,
            description: "供給形態",
          }
        forkOriginUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: "フォーク元 URL (INTERNAL_FORK の場合)",
          }
        createdAt: { type: string, format: date-time, description: "作成日時" }
        updatedAt: { type: string, format: date-time, description: "更新日時" }
      required:
        [
          id,
          ossId,
          version,
          modified,
          reviewStatus,
          scopeStatus,
          createdAt,
          updatedAt,
        ]
    OssVersionCreateRequest:
      type: object
      description: バージョン作成リクエスト
      required: [version]
      properties:
        version: { type: string, description: "バージョン文字列" }
        releaseDate:
          {
            type: string,
            format: date,
            nullable: true,
            description: "リリース日",
          }
        licenseExpressionRaw:
          { type: string, nullable: true, description: "生ライセンス式" }
        purl: { type: string, nullable: true, description: "package-url" }
        cpeList:
          type: array
          description: CPE 配列
          items: { type: string }
        hashSha256:
          type: string
          pattern: "^[A-Fa-f0-9]{64}$"
          nullable: true
          description: "アーカイブ SHA-256"
        modified: { type: boolean, default: false, description: "社内改変有無" }
        modificationDescription:
          { type: string, nullable: true, description: "改変概要" }
        supplierType:
          {
            $ref: "#/components/schemas/SupplierType",
            nullable: true,
            description: "供給形態",
          }
        forkOriginUrl:
          {
            type: string,
            format: uri,
            nullable: true,
            description: "フォーク元 URL",
          }
...
    PagedResult_OssComponent:
      type: object
      description: OSSコンポーネントページング結果
      properties:
        items:
          type: array
          description: 結果アイテム配列
          items: { $ref: "#/components/schemas/OssComponent" }
        page: { type: integer, description: "現在ページ (1 始まり)" }
        size: { type: integer, description: "ページサイズ" }
        total: { type: integer, description: "総件数" }
    PagedResult_OssVersion:
      type: object
      description: OSS バージョンページング結果
      properties:
        items:
          type: array
          description: 結果アイテム配列
          items: { $ref: "#/components/schemas/OssVersion" }
        page: { type: integer, description: "現在ページ" }
        size: { type: integer, description: "ページサイズ" }
        total: { type: integer, description: "総件数" }
...
    Role:
      type: string
      enum: [ADMIN, EDITOR, VIEWER]
      description: アクセス制御ロール（ADMIN=全権, EDITOR=編集可, VIEWER=参照のみ）
      x-enumDescriptions:
        - 管理者。ユーザー管理/ポリシー変更/全CRUDが可能
        - 編集者。OSS/プロジェクト/利用情報の作成・更新・削除が可能（ユーザー管理不可）
        - 閲覧者。読み取りのみ可能
    User:
      type: object
      description: システム利用ユーザー情報
      properties:
        id: { type: string, format: uuid, description: "ユーザーID" }
        username: { type: string, description: "ログインID（ユニーク）" }
        displayName: { type: string, description: "表示名" }
        email:
          {
            type: string,
            format: email,
            nullable: true,
            description: "メールアドレス",
          }
        roles:
          type: array
          description: 付与ロール一覧
          items: { $ref: "#/components/schemas/Role" }
        active: { type: boolean, description: "有効/無効フラグ" }
        createdAt: { type: string, format: date-time, description: "作成日時" }
        updatedAt: { type: string, format: date-time, description: "更新日時" }
      required: [id, username, roles, active, createdAt, updatedAt]
...
paths:
...
  /me:
    get:
      tags: [Users]
      summary: 現在ログイン中ユーザー情報取得
      operationId: getCurrentUser
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  ################################
  # 既存エンドポイント（以下は元定義を踏襲、必要なら401/403を追記）
  ################################
  /oss:
    get:
      tags: [OSS]
      summary: OSSコンポーネント一覧取得
      operationId: listOssComponents
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
        - name: name
          in: query
          schema: { type: string }
          description: 部分一致
        - name: layers
          in: query
          description: カンマ区切り Layer フィルタ (例 LIB,DB)
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
          description: タグ名 (正確一致)
        - name: inScopeOnly
          in: query
          schema: { type: boolean }
          description: true の場合 IN_SCOPE のバージョンを一つ以上持つもののみ
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedResult_OssComponent" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [OSS]
      summary: OSSコンポーネント作成
      operationId: createOssComponent
      x-rolesAllowed: [EDITOR, ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OssComponentCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OssComponent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /oss/{ossId}:
    get:
      tags: [OSS]
      summary: OSSコンポーネント詳細
      operationId: getOssComponent
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OssComponent" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [OSS]
      summary: OSSコンポーネント更新 (部分)
      operationId: updateOssComponent
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OssComponentUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OssComponent" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [OSS]
      summary: OSSコンポーネントを非推奨 (deprecated=true) に設定
      operationId: deprecateOssComponent
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /oss/{ossId}/versions:
    get:
      tags: [OSS Versions]
      summary: 指定 OSS のバージョン一覧
      operationId: listOssVersions
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - name: reviewStatus
          in: query
          schema: { $ref: "#/components/schemas/ReviewStatus" }
        - name: scopeStatus
          in: query
          schema: { $ref: "#/components/schemas/ScopeStatus" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PagedResult_OssVersion" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [OSS Versions]
      summary: バージョン追加
      operationId: createOssVersion
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OssVersionCreateRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OssVersion" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /oss/{ossId}/versions/{versionId}:
    get:
      tags: [OSS Versions]
      summary: バージョン詳細
      operationId: getOssVersion
      x-rolesAllowed: [VIEWER, EDITOR, ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: versionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OssVersion" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [OSS Versions]
      summary: バージョン更新
      operationId: updateOssVersion
      x-rolesAllowed: [EDITOR, ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: versionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OssVersionUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OssVersion" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [OSS Versions]
      summary: バージョン削除 (論理/物理は実装方針による)
      operationId: deleteOssVersion
      x-rolesAllowed: [ADMIN]
      parameters:
        - name: ossId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: versionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: No Content }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...